backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
pyjwt==2.8.0
cryptography==42.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.6.0

# Production
slowapi==0.1.9  # Rate limiting
python-json-logger==2.0.7  # Structured logging
prometheus-fastapi-instrumentator==6.1.0  # Metrics
